+++ src/app/app-routing.module.ts
@@ -27,7 +27,6 @@ import { DocsComponent } from './docs/docs.component';
 import { CanDeactivateDashboard } from './guards/can-deactivate-dashboard.guard';
 import { WarningsModuleComponent } from './dashboard/warnings-module/warnings-module.component';
 import { TimersModuleComponent } from './dashboard/timers-module/timers-module.component';
-import { BadgesComponent } from './dashboard/badges/badges.component';
 import { CratesComponent } from './crates/crates.component';
 import { LoginComponent } from './login/login.component';
 import { ReactionRolesModuleComponent } from './dashboard/reaction-roles-module/reaction-roles-module.component';
@@ -44,7 +43,6 @@ const routes: Routes = [
   { path: 'leaderboard/:id', component: LeaderboardModuleComponent, canActivate: [LeaderboardAuthGuard] },
 
   { path: 'dashboard', component: DashboardComponent, canActivate: [DashboardAuthGuard] },
-  { path: 'dashboard/badges', component: BadgesComponent, canActivate: [DashboardAuthGuard] },
   { path: 'dashboard/crates', component: CratesComponent, canActivate: [DashboardAuthGuard] },
   { path: 'dashboard/xp-card', component: XPCardComponent, canActivate: [DashboardAuthGuard] },