+++ src/app/guards/dashboard-auth.guard.ts
@@ -9,18 +9,14 @@ import { GuildService } from '../services/guild.service';
 export class DashboardAuthGuard implements CanActivate {
   constructor(
     private userService: UserService,
-    private guildService: GuildService,
     private router: Router) {}
 
   async canActivate() {
-    if (!this.userService.user)
-      await this.userService.updateUser();
-    if (!this.userService.savedUser)
-      await this.userService.updateSavedUser();
+    await this.userService.init();
     
     const canActivate = Boolean(this.userService.user);
     if (!canActivate)
-      this.router.navigate(['/login']);
+      this.router.navigate(['/']);
     return canActivate;
   }  
 }