+++ src/app/messages-widget/messages-widget.component.ts
@@ -15,7 +15,7 @@ export class MessagesWidgetComponent implements OnInit {
   chartReady = false;
   
   barChartOptions: ChartOptions = {
-    responsive: false,
+    responsive: true,
     scales: {
       yAxes: [{ display: false }],
       xAxes: [{ display: false }]