+++ src/environments/environment.prod.ts
@@ -1,12 +1,14 @@
 export const environment = {
-  production: true,
-  endpoint: '/api',
-  url: 'https://3pg.xyz',
   discordURL: 'https://discord.com/invite/xRT6Fz5',
+  githubURL: 'https://github.com/theADAMJR/2PG',
+  endpoint: '/api',
+  production: true,
   stripePublicKey: 'pk_live_kyJ00K40kgFhHMYzO7Bc9YjV00sVDW6IlV',
+  url: 'https://3pg.xyz',
   version: 'v2.2.0b',
   voteURLs: [
-    { name: '12 hours', url: '//top.gg/bot/525935335918665760/vote' },
-    { name: '24 hours', url: '//discordbotlist.com/bots/525935335918665760/upvote' }
+    { name: '12 hours (1)', url: '//dbots.co/bots/525935335918665760/vote' },
+    { name: '12 hours (2)', url: '//top.gg/bot/525935335918665760/vote' },
+    { name: '24 hours (3)', url: '//discordbotlist.com/bots/525935335918665760/upvote' }
   ]
 };