+++ src/environments/environment.ts
@@ -3,17 +3,15 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  endpoint: 'http://localhost:3000/api',
-  url: 'http://localhost:4200',
   discordURL: 'https://discord.com/invite/xRT6Fz5',
+  endpoint: 'http://localhost:3000/api',
+  githubURL: 'https://github.com/theADAMJR/2PG',
+  production: false,
   stripePublicKey: 'pk_test_5D57doatt7FpFolXlTHy8DvZ00TgL3h1WJ',
+  url: 'http://localhost:4200',
   version: 'v2.2.0b',
-  test: {
-    guildId: '598565371162656788'
-  },
   voteURLs: [
-    // { name: '12 hours (1)', url: '//dbots.co/bots/525935335918665760/vote' },
+    { name: '12 hours (1)', url: '//dbots.co/bots/525935335918665760/vote' },
     { name: '12 hours (2)', url: '//top.gg/bot/525935335918665760/vote' },
     { name: '24 hours (3)', url: '//discordbotlist.com/bots/525935335918665760/upvote' }
   ]